+++ src/main/java/br/gov/sp/fatec/mapskills/domain/theme/GameThemeRepository.java
@@ -8,8 +8,8 @@ package br.gov.sp.fatec.mapskills.domain.theme;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface GameThemeRepository extends CrudRepository<GameTheme, Integer> {
+public interface GameThemeRepository extends CrudRepository<GameTheme, Long> {
 	
-	public GameTheme findById(final int id);
+	public GameTheme findById(final long id);
 
 }
